requirements.txt

pandas
pillow
